Addressed feedback
Also swapped test file names Functional<-> Unit for RunCompletionHandler.
