+++ service/src/main/java/bio/terra/cbas/runsets/results/RunCompletionResult.java
@@ -5,7 +5,7 @@ import org.springframework.http.HttpStatus;
 public enum RunCompletionResult {
   SUCCESS("UPDATED_SUCCESSFULLY"),
   ERROR("RETRIABLE_ERROR"),
-  VALIDATION("VALIDATION_ERROR");
+  VALIDATION_ERROR("VALIDATION_ERROR");
   private final String value;
 
   RunCompletionResult(String value) {
@@ -20,7 +20,7 @@ public enum RunCompletionResult {
   public HttpStatus toHttpStatus() {
     if (SUCCESS.toString().equalsIgnoreCase(this.value)) {
       return HttpStatus.OK;
-    } else if (VALIDATION.toString().equalsIgnoreCase(this.value)) {
+    } else if (VALIDATION_ERROR.toString().equalsIgnoreCase(this.value)) {
       return HttpStatus.BAD_REQUEST;
     } else {
       return HttpStatus.INTERNAL_SERVER_ERROR;
