+++ service/src/test/java/bio/terra/cbas/runsets/results/TestRunCompletionHandlerUnit.java
@@ -1,30 +1,24 @@
 package bio.terra.cbas.runsets.results;
 
-import static bio.terra.cbas.models.CbasRunStatus.CANCELED;
-import static bio.terra.cbas.models.CbasRunStatus.COMPLETE;
 import static bio.terra.cbas.models.CbasRunStatus.RUNNING;
-import static bio.terra.cbas.models.CbasRunStatus.SYSTEM_ERROR;
+import static bio.terra.cbas.models.CbasRunStatus.UNKNOWN;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.ArgumentMatchers.isA;
-import static org.mockito.Mockito.any;
-import static org.mockito.Mockito.doThrow;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 
-import bio.terra.cbas.common.DateUtils;
+import bio.terra.cbas.common.exceptions.OutputProcessingException;
 import bio.terra.cbas.dao.RunDao;
 import bio.terra.cbas.dependencies.wds.WdsService;
-import bio.terra.cbas.dependencies.wds.WdsServiceException;
 import bio.terra.cbas.models.CbasRunSetStatus;
-import bio.terra.cbas.models.CbasRunStatus;
+import bio.terra.cbas.models.Method;
 import bio.terra.cbas.models.MethodVersion;
 import bio.terra.cbas.models.Run;
 import bio.terra.cbas.models.RunSet;
+import bio.terra.cbas.runsets.types.CoercionException;
 import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.util.StdDateFormat;
 import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
@@ -33,16 +27,29 @@ import com.fasterxml.jackson.module.paramnames.ParameterNamesModule;
 import com.google.gson.Gson;
 import cromwell.client.model.RunLog;
 import java.time.OffsetDateTime;
-import java.util.Collections;
-import java.util.List;
 import java.util.UUID;
+import org.databiosphere.workspacedata.model.RecordAttributes;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.mockito.ArgumentCaptor;
 import org.springframework.test.context.ContextConfiguration;
 
 @ContextConfiguration(classes = RunCompletionHandler.class)
 class TestRunCompletionHandlerUnit {
+
+  private static final OffsetDateTime methodCreatedTime = OffsetDateTime.now();
+  private static final OffsetDateTime runSubmittedTime = OffsetDateTime.now();
+  private static final UUID runningRunId1 = UUID.randomUUID();
+  private static final String runningRunEngineId1 = UUID.randomUUID().toString();
+  private static final String runningRunEntityId1 = UUID.randomUUID().toString();
+  // Set the last status update a few minutes ago to avoid tripping over the "recent poll
+  // ineligibility" requirement:
+  private static final OffsetDateTime runningRunStatusUpdateTime =
+      OffsetDateTime.now().minusMinutes(5);
+  private static final UUID runningRunId2 = UUID.randomUUID();
+  private static final String runningRunEngineId2 = UUID.randomUUID().toString();
+  private static final String runningRunEntityId2 = UUID.randomUUID().toString();
+  private static final String errorMessages = null;
 
   public ObjectMapper objectMapper =
       new ObjectMapper()
@@ -51,9 +58,7 @@ class TestRunCompletionHandlerUnit {
           .registerModule(new JavaTimeModule())
           .setDateFormat(new StdDateFormat())
           .setDefaultPropertyInclusion(JsonInclude.Include.NON_ABSENT);
-
-  private RunDao runDao;
-  private WdsService wdsService;
+  private RunCompletionHandler runCompletionHandler;
   static String outputDefinition =
       """
         [
@@ -64,435 +69,179 @@ class TestRunCompletionHandlerUnit {
           }
         ]
       """;
+  private static final UUID runSetId = UUID.randomUUID();
+  private static final RunSet runSet =
+      new RunSet(
+          runSetId,
+          new MethodVersion(
+              UUID.randomUUID(),
+              new Method(
+                  UUID.randomUUID(),
+                  "methodName",
+                  "methodDescription",
+                  methodCreatedTime,
+                  runSetId,
+                  "method source"),
+              "version name",
+              "version description",
+              methodCreatedTime,
+              runSetId,
+              "file:///method/source/url"),
+          "runSetName",
+          "runSetDescription",
+          true,
+          false,
+          CbasRunSetStatus.UNKNOWN,
+          runSubmittedTime,
+          runSubmittedTime,
+          runSubmittedTime,
+          0,
+          0,
+          "inputDefinition",
+          outputDefinition,
+          "entityType",
+          "user-foo");
 
-  static String outputs =
-      """
+  private static final RunSet runSetNoOutputs =
+      new RunSet(
+          runSetId,
+          new MethodVersion(
+              UUID.randomUUID(),
+              new Method(
+                  UUID.randomUUID(),
+                  "methodName",
+                  "methodDescription",
+                  methodCreatedTime,
+                  runSetId,
+                  "method source"),
+              "version name",
+              "version description",
+              methodCreatedTime,
+              runSetId,
+              "file:///method/source/url"),
+          "runSetName",
+          "runSetDescription",
+          true,
+          false,
+          CbasRunSetStatus.UNKNOWN,
+          runSubmittedTime,
+          runSubmittedTime,
+          runSubmittedTime,
+          0,
+          0,
+          "inputDefinition",
+          "[]",
+          "entityType",
+          "user-foo");
+
+  final Run runToUpdate1 =
+      new Run(
+          runningRunId1,
+          runningRunEngineId1,
+          runSet,
+          runningRunEntityId1,
+          runSubmittedTime,
+          RUNNING,
+          runningRunStatusUpdateTime,
+          runningRunStatusUpdateTime,
+          errorMessages);
+
+  final Run runToUpdate2 =
+      new Run(
+          runningRunId2,
+          runningRunEngineId2,
+          runSetNoOutputs,
+          runningRunEntityId2,
+          runSubmittedTime,
+          UNKNOWN,
+          runningRunStatusUpdateTime,
+          runningRunStatusUpdateTime,
+          errorMessages);
+
+  @BeforeEach
+  public void init() {
+    RunDao runsDao = mock(RunDao.class);
+    WdsService wdsService = mock(WdsService.class);
+    runCompletionHandler = new RunCompletionHandler(runsDao, wdsService, objectMapper);
+  }
+
+  @Test
+  void buildRecordAttributesFromWorkflowOutputsSucceeds()
+      throws OutputProcessingException, CoercionException, JsonProcessingException {
+    // Using Gson here since Cromwell client uses it to interpret runLogValue into Java objects.
+    Gson object = new Gson();
+    String rawOutputs =
+        """
         {
             "outputs": {
               "wf_hello.hello.salutation": "Hello batch!"
             }
-        }
+          }
         """;
 
-  static String outputsToThrow =
-      """
+    Object cromwellOutputs = object.fromJson(rawOutputs, RunLog.class).getOutputs();
+
+    RecordAttributes mockAttributes = new RecordAttributes();
+    mockAttributes.put("foo_name", "Hello batch!");
+
+    var recordAttributes =
+        runCompletionHandler.buildRecordAttributesFromWorkflowOutputs(
+            runToUpdate1, cromwellOutputs);
+
+    assertEquals(mockAttributes, recordAttributes);
+  }
+
+  @Test
+  void buildRecordAttributesFromWorkflowOutputsReturnNull()
+      throws OutputProcessingException, CoercionException, JsonProcessingException {
+    // Using Gson here since Cromwell client uses it to interpret runLogValue into Java objects.
+    Gson object = new Gson();
+    String rawOutputs =
+        """
         {
             "outputs": {
-              "wf_hello.hello.salutations": "Hello batch!"
+              "wf_hello.hello.salutation": "Hello batch!"
             }
-        }
+          }
         """;
-  static String emptyOutputs = """
+    Object cromwellOutputs = object.fromJson(rawOutputs, RunLog.class).getOutputs();
+
+    var recordAttributes =
+        runCompletionHandler.buildRecordAttributesFromWorkflowOutputs(
+            runToUpdate2, cromwellOutputs);
+
+    assertNull(recordAttributes);
+  }
+
+  @Test
+  void buildRecordAttributesFromWorkflowOutputsThrowsMisspelledOutputs() {
+    // Using Gson here since Cromwell client uses it to interpret runLogValue into Java objects.
+    Gson object = new Gson();
+    String misspelledRawOutputs =
+        """
         {
-            "outputs":{}
-        }
+            "outputs": {
+              "wf_hello.hello.salutationz": "Hello batch!"
+            }
+          }
         """;
 
-  @BeforeEach
-  void init() {
-    runDao = mock(RunDao.class);
-    wdsService = mock(WdsService.class);
+    Object cromwellOutputs = object.fromJson(misspelledRawOutputs, RunLog.class).getOutputs();
+
+    Assertions.assertThrows(
+        OutputProcessingException.class,
+        () ->
+            runCompletionHandler.buildRecordAttributesFromWorkflowOutputs(
+                runToUpdate1, cromwellOutputs));
   }
 
   @Test
-  void updateRunCompletionSucceededNoOutputsComplete() {
-    RunCompletionHandler runCompletionHandler =
-        new RunCompletionHandler(runDao, wdsService, objectMapper);
-
-    UUID runId1 = UUID.randomUUID();
-    RunSet runSet1 = createRunSet(UUID.randomUUID(), "[]");
-    Run run1Incomplete = createTestRun(runId1, runSet1, RUNNING);
-
-    // Set up mocks:
-    when(runDao.getRuns(any())).thenReturn(List.of(run1Incomplete));
-    when(runDao.updateRunStatus(eq(runId1), eq(CbasRunStatus.COMPLETE), isA(OffsetDateTime.class)))
-        .thenReturn(1);
-
-    // Run the results update:
-    var result =
-        runCompletionHandler.updateResults(run1Incomplete, COMPLETE, null, Collections.emptyList());
-
-    // Validate the results:
-    verify(runDao, times(1)).updateRunStatus(eq(runId1), eq(COMPLETE), any());
-    verify(runDao, times(0)).updateRunStatusWithError(eq(runId1), eq(COMPLETE), any(), anyString());
-    assertEquals(RunCompletionResult.SUCCESS, result);
+  void hasOutputDefinitionReturnsTrue() throws JsonProcessingException {
+    assertTrue(runCompletionHandler.hasOutputDefinition(runToUpdate1));
   }
 
   @Test
-  void updateRunCompletionWorkflowErrorsRecordedDateTime() {
-    RunCompletionHandler runCompletionHandler =
-        new RunCompletionHandler(runDao, wdsService, objectMapper);
-    var errorList = List.of("error1", "error 2");
-    UUID runId1 = UUID.randomUUID();
-    Run run1Incomplete = createTestRun(runId1, null, SYSTEM_ERROR);
-
-    // Set up mocks:
-    when(runDao.updateRunStatusWithError(eq(runId1), eq(SYSTEM_ERROR), any(), any())).thenReturn(1);
-
-    // Run the results update:
-    var result = runCompletionHandler.updateResults(run1Incomplete, SYSTEM_ERROR, null, errorList);
-    ArgumentCaptor<String> errors = ArgumentCaptor.forClass(String.class);
-    // Validate the results:
-    verify(runDao, times(0)).updateRunStatus(eq(runId1), any(), any());
-    verify(runDao, times(1)).updateRunStatusWithError(eq(runId1), any(), any(), errors.capture());
-    verify(runDao, times(0)).updateLastPolledTimestamp(runId1);
-    assertEquals("error1, error 2", errors.getValue());
-    assertEquals(RunCompletionResult.SUCCESS, result);
-  }
-
-  @Test
-  void updateRunCompletionNoStatusChangeNoOutputsUpdateDateTimeNoRecord() {
-    RunCompletionHandler runCompletionHandler =
-        new RunCompletionHandler(runDao, wdsService, objectMapper);
-
-    UUID runId1 = UUID.randomUUID();
-    Run run1Incomplete = createTestRun(runId1, null, SYSTEM_ERROR);
-
-    // Set up mocks:
-    when(runDao.updateLastPolledTimestamp(runId1)).thenReturn(0);
-
-    // Run the results update:
-    var result = runCompletionHandler.updateResults(run1Incomplete, SYSTEM_ERROR, null, null);
-
-    // Validate the results:
-    verify(runDao, times(0)).updateRunStatus(eq(runId1), any(), any());
-    verify(runDao, times(0)).updateRunStatusWithError(eq(runId1), any(), any(), anyString());
-    verify(runDao, times(1)).updateLastPolledTimestamp(runId1);
-    assertEquals(RunCompletionResult.ERROR, result);
-  }
-
-  @Test
-  void updateRunCompletionSucceededNoOutputsNoErrorsToUpdate() throws WdsServiceException {
-    RunCompletionHandler runCompletionHandler =
-        new RunCompletionHandler(runDao, wdsService, objectMapper);
-    UUID runId1 = UUID.randomUUID();
-    RunSet runSet1 = createRunSet(UUID.randomUUID(), "[]");
-    Run run1Incomplete = createTestRun(runId1, runSet1, RUNNING);
-
-    // Set up mocks:
-    when(runDao.updateRunStatus(eq(runId1), eq(CbasRunStatus.CANCELED), isA(OffsetDateTime.class)))
-        .thenReturn(1);
-    // Run the results update:
-    var result =
-        runCompletionHandler.updateResults(
-            run1Incomplete, CbasRunStatus.CANCELED, null, Collections.emptyList());
-
-    // Validate the results:
-    verify(wdsService, times(0)).updateRecord(any(), any(), any());
-    verify(runDao, times(1)).updateRunStatus(eq(runId1), eq(CANCELED), any());
-    verify(runDao, times(0)).updateRunStatusWithError(eq(runId1), any(), any(), anyString());
-    assertEquals(RunCompletionResult.SUCCESS, result);
-  }
-
-  @Test
-  void updateRunCompletionSucceededNoStatusUpdate() throws WdsServiceException {
-    RunCompletionHandler runCompletionHandler =
-        new RunCompletionHandler(runDao, wdsService, objectMapper);
-    UUID runId1 = UUID.randomUUID();
-    RunSet runSet1 = createRunSet(UUID.randomUUID(), outputDefinition);
-    Run run1Incomplete = createTestRun(runId1, runSet1, CANCELED);
-
-    // Set up mocks:
-    when(runDao.updateLastPolledTimestamp(runId1)).thenReturn(1);
-    // Run the results update:
-    var result =
-        runCompletionHandler.updateResults(
-            run1Incomplete, CbasRunStatus.CANCELED, "[]", Collections.emptyList());
-
-    // Validate the results:
-    verify(wdsService, times(0)).updateRecord(any(), any(), any());
-    verify(runDao, times(1)).updateLastPolledTimestamp(runId1);
-    verify(runDao, times(0)).updateRunStatus(eq(runId1), any(), any());
-    verify(runDao, times(0)).updateRunStatusWithError(eq(runId1), any(), any(), anyString());
-    assertEquals(RunCompletionResult.SUCCESS, result);
-  }
-
-  @Test
-  void updateRunCompletionFailedNoRecordsUpdated() {
-    RunCompletionHandler runCompletionHandler =
-        new RunCompletionHandler(runDao, wdsService, objectMapper);
-    UUID runId1 = UUID.randomUUID();
-    RunSet runSet1 = createRunSet(UUID.randomUUID(), "[]");
-    Run run1Incomplete = createTestRun(runId1, runSet1, RUNNING);
-
-    // Set up mocks:
-    when(runDao.updateRunStatus(eq(runId1), eq(CbasRunStatus.CANCELED), isA(OffsetDateTime.class)))
-        .thenReturn(0);
-    // Run the results update:
-    var result =
-        runCompletionHandler.updateResults(
-            run1Incomplete, CbasRunStatus.CANCELED, null, Collections.emptyList());
-
-    // Validate the results:
-    verify(runDao, times(1)).updateRunStatus(eq(runId1), eq(CANCELED), any());
-    verify(runDao, times(0)).updateRunStatusWithError(eq(runId1), any(), any(), anyString());
-    assertEquals(RunCompletionResult.ERROR, result);
-    assertEquals(RUNNING, run1Incomplete.status());
-  }
-
-  @Test
-  void updateRunCompletionSucceededWithOutputsSaved() throws WdsServiceException {
-    RunCompletionHandler runCompletionHandler =
-        new RunCompletionHandler(runDao, wdsService, objectMapper);
-    // Set up run to expect non-empty outputs
-    RunSet runSet1 = createRunSet(UUID.randomUUID(), outputDefinition);
-    UUID runId1 = UUID.randomUUID();
-    Run run1Incomplete = createTestRun(runId1, runSet1, RUNNING);
-    // Using Gson here since Cromwell client uses it to interpret runLogValue into Java objects.
-    Gson object = new Gson();
-    Object cromwellOutputs = object.fromJson(outputs, RunLog.class).getOutputs();
-
-    // Set up mocks:
-    when(runDao.updateRunStatus(eq(runId1), eq(CbasRunStatus.COMPLETE), isA(OffsetDateTime.class)))
-        .thenReturn(1);
-    // Run the results update:
-    var result =
-        runCompletionHandler.updateResults(run1Incomplete, COMPLETE, cromwellOutputs, null);
-
-    // Validate the results:
-    verify(runDao, times(1)).updateRunStatus(eq(runId1), eq(COMPLETE), any());
-    verify(runDao, times(0)).updateRunStatusWithError(eq(runId1), any(), any(), anyString());
-    verify(wdsService, times(1)).updateRecord(any(), any(), any());
-    assertEquals(RunCompletionResult.SUCCESS, result);
-  }
-
-  @Test
-  void updateRunCompletionSucceededWithEmptyOutputs() throws WdsServiceException {
-    RunCompletionHandler runCompletionHandler =
-        new RunCompletionHandler(runDao, wdsService, objectMapper);
-    // Set up run to expect non-empty outputs
-    RunSet runSet = createRunSet(UUID.randomUUID(), "[]");
-    UUID runId1 = UUID.randomUUID();
-    Run run1Incomplete = createTestRun(runId1, runSet, RUNNING);
-
-    // Using Gson here since Cromwell client uses it to interpret runLogValue into Java objects.
-    Gson object = new Gson();
-    Object cromwellOutputs = object.fromJson(emptyOutputs, RunLog.class).getOutputs();
-
-    // Set up mocks:
-    when(runDao.updateRunStatus(eq(runId1), eq(CbasRunStatus.COMPLETE), isA(OffsetDateTime.class)))
-        .thenReturn(1);
-    // Run the results update:
-    var result =
-        runCompletionHandler.updateResults(run1Incomplete, COMPLETE, cromwellOutputs, null);
-    // Validate the results:
-    verify(runDao, times(1)).updateRunStatus(eq(runId1), any(), any());
-    verify(runDao, times(0))
-        .updateRunStatusWithError(eq(runId1), eq(SYSTEM_ERROR), any(), anyString());
-    verify(wdsService, times(0)).updateRecord(any(), any(), any());
-
-    assertEquals(RunCompletionResult.SUCCESS, result);
-  }
-
-  @Test
-  void updateRunCompletionSucceededWhenWdsThrowsSavedStatusWithErrors() throws WdsServiceException {
-    RunCompletionHandler runCompletionHandler =
-        new RunCompletionHandler(runDao, wdsService, objectMapper);
-    // Set up run to expect non-empty outputs
-    RunSet runSet = createRunSet(UUID.randomUUID(), outputDefinition);
-    UUID runId1 = UUID.randomUUID();
-    Run run1Incomplete = createTestRun(runId1, runSet, RUNNING);
-    // Using Gson here since Cromwell client uses it to interpret runLogValue into Java objects.
-    Gson object = new Gson();
-    Object cromwellOutputs = object.fromJson(outputs, RunLog.class).getOutputs();
-    List<String> failures = createWorkflowErrorsList();
-
-    // Set up mocks:
-    when(runDao.updateRunStatusWithError(eq(runId1), eq(SYSTEM_ERROR), any(), anyString()))
-        .thenReturn(1);
-    // WDS throws
-    doThrow(new RuntimeException("Some WDS error"))
-        .when(wdsService)
-        .updateRecord(any(), any(), any());
-    // Run the results update:
-    var result =
-        runCompletionHandler.updateResults(run1Incomplete, COMPLETE, cromwellOutputs, failures);
-
-    // Validate the results:
-    verify(runDao, times(0)).updateRunStatus(eq(runId1), any(), any());
-    verify(runDao, times(1))
-        .updateRunStatusWithError(eq(runId1), eq(SYSTEM_ERROR), any(), anyString());
-
-    assertEquals(RunCompletionResult.SUCCESS, result);
-  }
-
-  @Test
-  void updateRunCompletionReturnsValidationWithOutputsErrorProcessing() throws WdsServiceException {
-    RunCompletionHandler runCompletionHandler =
-        new RunCompletionHandler(runDao, wdsService, objectMapper);
-    // Set up run to expect non-empty outputs
-    RunSet runSet = createRunSet(UUID.randomUUID(), outputDefinition);
-    UUID runId1 = UUID.randomUUID();
-    Run run1Incomplete = createTestRun(runId1, runSet, RUNNING);
-    // Using Gson here since Cromwell client uses it to interpret runLogValue into Java objects.
-    Gson object = new Gson();
-    Object cromwellOutputs = object.fromJson(outputsToThrow, RunLog.class).getOutputs();
-    List<String> failures = createWorkflowErrorsList();
-
-    // Set up mocks:
-    when(runDao.updateRunStatusWithError(eq(runId1), eq(SYSTEM_ERROR), any(), anyString()))
-        .thenReturn(1);
-
-    // Run the results update:
-    var result =
-        runCompletionHandler.updateResults(run1Incomplete, COMPLETE, cromwellOutputs, failures);
-
-    // Validate the results:
-    verify(wdsService, times(0)).updateRecord(any(), any(), any());
-    verify(runDao, times(0)).updateRunStatus(eq(runId1), any(), any());
-    verify(runDao, times(0))
-        .updateRunStatusWithError(eq(runId1), eq(SYSTEM_ERROR), any(), anyString());
-
-    assertEquals(RunCompletionResult.VALIDATION, result);
-  }
-
-  @Test
-  void updateRunCompletionSucceededWithEmptyOutputsNoFailures() {
-    RunCompletionHandler runCompletionHandler =
-        new RunCompletionHandler(runDao, wdsService, objectMapper);
-
-    // Using Gson here since Cromwell client uses it to interpret runLogValue into Java objects.
-    Gson object = new Gson();
-    Object cromwellOutputs = object.fromJson(emptyOutputs, RunLog.class).getOutputs();
-
-    UUID runId1 = UUID.randomUUID();
-    Run run1Incomplete = createTestRun(runId1, null, RUNNING);
-
-    // Set up mocks:
-    when(runDao.updateRunStatus(eq(runId1), eq(SYSTEM_ERROR), any())).thenReturn(1);
-
-    // Run the results update:
-    var result =
-        runCompletionHandler.updateResults(run1Incomplete, SYSTEM_ERROR, cromwellOutputs, null);
-
-    // Validate the results:
-    // Since error count is 0, only Status is going to be updated in DB.
-    verify(runDao, times(1)).updateRunStatus(eq(runId1), eq(SYSTEM_ERROR), any());
-    verify(runDao, times(0)).updateRunStatusWithError(eq(runId1), any(), any(), anyString());
-    assertEquals(RunCompletionResult.SUCCESS, result);
-  }
-
-  @Test
-  void updateRunCompletionSuccessSavingFailures() {
-    RunCompletionHandler runCompletionHandler =
-        new RunCompletionHandler(runDao, wdsService, objectMapper);
-
-    // Using Gson here since Cromwell client uses it to interpret runLogValue into Java objects.
-    Gson object = new Gson();
-    Object cromwellOutputs = object.fromJson(emptyOutputs, RunLog.class).getOutputs();
-
-    UUID runId1 = UUID.randomUUID();
-    Run run1Incomplete = createTestRun(runId1, null, RUNNING);
-
-    // Set up mocks:
-    when(runDao.updateRunStatusWithError(eq(runId1), eq(SYSTEM_ERROR), any(), anyString()))
-        .thenReturn(1);
-    List<String> errors = createWorkflowErrorsList();
-
-    // Run the results update:
-    var result =
-        runCompletionHandler.updateResults(run1Incomplete, SYSTEM_ERROR, cromwellOutputs, errors);
-
-    // Validate the results:
-    verify(runDao, times(0)).updateRunStatus(eq(runId1), any(), any());
-    verify(runDao, times(1))
-        .updateRunStatusWithError(eq(runId1), eq(SYSTEM_ERROR), any(), anyString());
-    assertEquals(RunCompletionResult.SUCCESS, result);
-  }
-
-  @Test
-  void updateRunCompletionFailedErrorsPulledNoRecordUpdated() {
-    RunCompletionHandler runCompletionHandler =
-        new RunCompletionHandler(runDao, wdsService, objectMapper);
-
-    // Using Gson here since Cromwell client uses it to interpret runLogValue into Java objects.
-    Gson object = new Gson();
-    Object cromwellOutputs = object.fromJson(emptyOutputs, RunLog.class).getOutputs();
-
-    UUID runId1 = UUID.randomUUID();
-    Run run1Incomplete = createTestRun(runId1, null, RUNNING);
-
-    // Set up mocks:
-    when(runDao.updateRunStatusWithError(eq(runId1), eq(SYSTEM_ERROR), any(), anyString()))
-        .thenReturn(0);
-    List<String> errors = createWorkflowErrorsList();
-
-    // Run the results update:
-    var result =
-        runCompletionHandler.updateResults(run1Incomplete, SYSTEM_ERROR, cromwellOutputs, errors);
-
-    // Validate the results:
-    verify(runDao, times(0)).updateRunStatus(eq(runId1), any(), any());
-    verify(runDao, times(1))
-        .updateRunStatusWithError(eq(runId1), eq(SYSTEM_ERROR), any(), anyString());
-    assertEquals(RunCompletionResult.ERROR, result);
-  }
-
-  @Test
-  void updateRunCompletionSuccessWithEmptyFailures() {
-    RunCompletionHandler runCompletionHandler =
-        new RunCompletionHandler(runDao, wdsService, objectMapper);
-
-    // Using Gson here since Cromwell client uses it to interpret runLogValue into Java objects.
-    Gson object = new Gson();
-    Object cromwellOutputs = object.fromJson(emptyOutputs, RunLog.class).getOutputs();
-
-    UUID runId1 = UUID.randomUUID();
-    Run run1Incomplete = createTestRun(runId1, null, RUNNING);
-
-    // Set up mocks:
-    when(runDao.updateRunStatus(eq(runId1), eq(SYSTEM_ERROR), any())).thenReturn(1);
-    List<String> failures = Collections.emptyList();
-
-    // Run the results update:
-    var result =
-        runCompletionHandler.updateResults(run1Incomplete, SYSTEM_ERROR, cromwellOutputs, failures);
-
-    // Validate the results:
-    // Since error count is 0, only Status is going to be updated in DB.
-    verify(runDao, times(1)).updateRunStatus(eq(runId1), any(), any());
-    verify(runDao, times(0))
-        .updateRunStatusWithError(eq(runId1), eq(SYSTEM_ERROR), any(), anyString());
-    assertEquals(RunCompletionResult.SUCCESS, result);
-  }
-
-  private Run createTestRun(UUID runId, RunSet runSet, CbasRunStatus status) {
-
-    String engineId1 = "mockEngine1";
-    String recordId1 = "mockRecordId1";
-    OffsetDateTime submissionTimestamp1 = DateUtils.currentTimeInUTC();
-    return new Run(
-        runId, engineId1, runSet, recordId1, submissionTimestamp1, status, null, null, null);
-  }
-
-  private RunSet createRunSet(UUID runSetId, String outputDefinition) {
-    return new RunSet(
-        runSetId,
-        new MethodVersion(
-            UUID.randomUUID(),
-            null,
-            "version name",
-            "version description",
-            OffsetDateTime.now(),
-            runSetId,
-            "file:///method/source/url"),
-        "runSetName",
-        "runSetDescription",
-        true,
-        false,
-        CbasRunSetStatus.UNKNOWN,
-        OffsetDateTime.now(),
-        OffsetDateTime.now(),
-        OffsetDateTime.now(),
-        0,
-        0,
-        "inputDefinition",
-        outputDefinition,
-        "entityType",
-        "user-foo");
-  }
-
-  private List<String> createWorkflowErrorsList() {
-    return List.of("Workflow error1", "Workflow error2");
+  void hasOutputDefinitionReturnsFalse() throws JsonProcessingException {
+    assertFalse(runCompletionHandler.hasOutputDefinition(runToUpdate2));
   }
 }
